==> ST-breast/3-10x_breast_immune_reference.Rmd <==
---
author: "Marc Elosua-Bayes"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
    df_print: paged
editor_options: 
  chunk_output_type: console
params: 
    sample_id: "Default!"
    clust_vr: "Default!"
title: "`r sprintf('3-10x_breast_immune_reference %s', {params$sample_id})`"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message = FALSE, warning = FALSE)
options(width = 1200)
```

## Introduction
In this R markdown document we will map immune cells onto the breast cancer tissue. 

```{r}
library(Seurat)
library(tidyverse)
library(Matrix)
library(SPOTlight)
source(here::here("utils/bin.r"))
```

## Paths
```{r}
source(here::here("misc/paths.R"))

"{an_breast_10x}/{robj_dir}" %>%
  glue::glue() %>%
  here::here() %>%
  dir.create(
    path = ,
    showWarnings = FALSE,
    recursive = TRUE)

"{an_breast_10x}/{plt_dir}" %>%
  glue::glue() %>%
  here::here() %>%
  dir.create(
    path = ,
    showWarnings = FALSE,
    recursive = TRUE)
```

## Parameters
```{r}
set.seed(1243)

sample_id <- params$sample_id
# sample_id <- "breast_1"
# sample_id <- "breast_2"
clust_vr <- params$clust_vr
# clust_vr <- "lv1_annot"
# clust_vr <- "lv2_annot"

trn <- "melanoma"
cl_n <- 100
hvg <- 3000
ntop <- NULL
transf <- "uv"
method <- "nsNMF"
min_cont <- 0

if (is.null(ntop)) {
  spotlight_id <- glue::glue("trn-{trn}_cln-{cl_n}_hvg-{hvg}_ntop-NULL_transf-{transf}_method-{method}_mincont-{min_cont}")
} else {
  spotlight_id <- glue::glue("trn-{trn}_cln-{cl_n}_hvg-{hvg}_ntop-{ntop}_transf-{transf}_method-{method}_mincont-{min_cont}")
}
```

Define cell types of interest depending on the annotation level
```{r}
if (clust_vr == "lv2_annot") {
  ct_interest <- c("Plasma.B.cells", "T.regs", "T.helper.Th17",
                   "Proliferation", "CD8.exhausted",
                   "CD8.cytotoxic", "NK", "Macrophages.SPP1", "TAMs.C1QC")
} else if (clust_vr == "lv1_annot") {
  ct_interest <- c("Plasma B cells", "T cells regulatory", "T helper cells",
                   "Proliferation", "CD8 pre-exhausted",
                   "CD8 terminally exhausted", "NK", "Macrophages SPP1", "TAMs C1QC")
}
```


## Load data
Breast cancer data loaded here comes from 1-10x_breast_QC.Rmd
```{r}
### Spatial breast cancer
se_obj <- "{an_breast_10x}/{robj_dir}/qc_se_{sample_id}.rds" %>%
  glue::glue() %>%
  here::here() %>%
  readRDS(file = .)
```

Create a name/color dataframe
```{r}
# source(here::here("misc/col_df_scrpt.R"))
# source(here::here("misc/col_df_scrpt2.R"))

# col_vec <- readRDS(here::here("data/immune_reference/complete_cell_type_palette.rds"))
# col_df <- data.frame(col_vec) %>%
#   tibble::rownames_to_column("plt_name") %>%
#   dplyr::mutate(ct_name = stringr::str_replace_all(
#     string = plt_name,
#     pattern = "[[:punct:]]|[[:blank:]]",
#     replacement = "."))
# prloif_df <- data.frame(plt_name = "Proliferation", col_vec = "Brown", ct_name = "Proliferation")
# col_df <- rbind(col_df, prloif_df)
# col_df <- col_df %>%
#   dplyr::mutate(
#     # plt_name = dplyr::if_else(plt_name == "Th17 cells", "CD4 effector memory", plt_name),
#     # ct_name = dplyr::if_else(ct_name == "Th17.cells", "CD4.effector.memory", ct_name),
#     plt_name = dplyr::if_else(plt_name == "TAMs M2", "TAMs C1QC", plt_name)
#   )
# saveRDS(object = col_df, file = here::here("misc/col_df.rds"))
col_df <- readRDS(here::here("misc/col_df.rds"))
```

Subset ICA to just use cells from the melanoma subset since we have a good representation of cells for each cluster from this dataset.
```{r eval = FALSE}
ica_melanoma2_path <- "{an_breast_10x}/{robj_dir}/ica_melanoma2.rds" %>%
  glue::glue() %>%
  here::here()

if (file.exists(ica_melanoma2_path)) {
  
  ica_sub <- readRDS(file = ica_melanoma2_path)
} else {
  # ica_se <- readRDS("/scratch/devel/pnieto/TIL_Atlas/TICA/output/integrated_renamed_filtered.rds")
  ica_se <- "/scratch/devel/pnieto/TIL_Atlas/atlas_final_dataset/new_annot/TICAtlas.rds" %>%
    here::here() %>%
    readRDS(.)
  ### Immune reference atlas
  ica_sub <- subset(ica_se, subset = source == "melanoma2")
  rm(ica_se)
  
  saveRDS(object = ica_sub, file = ica_melanoma2_path)
}

table(ica_sub@meta.data$lv1_annot)
table(ica_sub@meta.data$lv2_annot)
```

Change variable names to remove non-standard characters
```{r eval = FALSE}
ica_sub[["specific_cell_type_mod"]] <- stringr::str_replace_all(
  string = as.character(ica_sub@meta.data[, clust_vr]),
  pattern = "[[:punct:]]|[[:blank:]]",
  replacement = ".")

prolif_vec <- c("T.cells.proliferative", "Macrophages.and.monocytes.proliferative",
                "B.cells.proliferative", "Macrophages.proliferative")

ica_sub[["specific_cell_type_mod"]] <- 
  dplyr::if_else(ica_sub@meta.data[, "specific_cell_type_mod"] %in% prolif_vec,
                 "Proliferation", ica_sub@meta.data[, "specific_cell_type_mod"])

table(ica_sub@meta.data$specific_cell_type_mod)
```

Immune cell marker genes
```{r eval = FALSE}
ica_sub <- Seurat::SCTransform(object = ica_sub, assay = "RNA")

Seurat::Idents(ica_sub) <- as.character(ica_sub@meta.data[, "specific_cell_type_mod"])

ica_markers <- Seurat::FindAllMarkers(object = ica_sub,
                       assay = "SCT",
                       slot = "data",
                       only.pos = TRUE,
                       logfc.threshold = 0,
                       min.pct = 0,
                       max.cells.per.ident = 500)


"data/immune_reference/ica_markers_melanoma2_{clust_vr}.rds" %>%
  glue::glue()%>%
  here::here( ) %>%
  saveRDS(
    object = ica_markers,
    file = .)
```

Load marker genes per cluster
```{r eval = FALSE}
ica_markers <- "data/immune_reference/ica_markers_melanoma2_{clust_vr}.rds" %>%
  glue::glue()%>%
  here::here() %>%
  readRDS(file = .)

# ica_markers <- readRDS(file = here::here("data/immune_reference/ica_markers_melanoma2.rds"))

# if (clust_vr == "lv1_annot") {
  # Remove MAST cells
ica_markers <- ica_markers %>% dplyr::filter(cluster != "Mast.cells")
# } else if (clust_vr == "lv2_annot") {
  # Remove MAST cells
  # ica_markers <- ica_markers %>% dplyr::filter(cluster != "Mast.cells")

# }
```

Look at the proliferation markers
```{r  eval = FALSE}
ica_markers %>%
  dplyr::filter(cluster == "Proliferation") %>%
  head(20)
```

## Deconvolution
Run deconvolution.
```{r message = FALSE,  eval = FALSE}
# Remove Mast cells
ica_sub <- subset(ica_sub, subset = specific_cell_type_mod != "Mast.cells")
# Create Seurat object with filtered gene matrix after removing Mast cells
ica_sub <- Seurat::CreateSeuratObject(
  counts = ica_sub@assays$RNA@counts[
    sparseMatrixStats::rowSums2(ica_sub@assays$RNA@counts) != 0, ],
  meta.data = ica_sub@meta.data)

# Remove empty spatial genes
# keep_genes <- rowSums(as.matrix(se_obj@assays$Spatial@counts)) != 0
# table(rowSums(ica_sub@assays$RNA@counts) != 0)

# Run deconvolution
decon_mtrx_ls <- SPOTlight::spotlight_deconvolution(
  se_sc = ica_sub,
  counts_spatial = se_obj@assays$Spatial@counts,
  clust_vr = "specific_cell_type_mod",
  cluster_markers = ica_markers,
  cl_n = cl_n,
  hvg = hvg,
  ntop = ntop,
  transf = transf,
  method = method,
  min_cont = min_cont,
  assay = "RNA",
  slot = "counts")

# saveRDS(object = decon_mtrx_ls,
#         file = sprintf("%s/%s/decon_mtrx_breast_cancer_10x_atlas_%s_reg.rds",
#                        an_breast_10x, robj_dir, spotlight_id))
# 
# saveRDS(object = decon_mtrx_ls,
#         file = sprintf("%s/%s/decon_mtrx_breast_cancer_10x_atlas_%s_latent.rds",
#                        an_breast_10x, robj_dir, spotlight_id))

"{an_breast_10x}/{robj_dir}/decon_mtrx_breast_cancer_10x_atlas_{sample_id}_{spotlight_id}_{clust_vr}.rds" %>%
  glue::glue() %>%
  here::here() %>%
  saveRDS(object = decon_mtrx_ls, file = .)
```

Add deconvolution matrix to Seurat object metadata
```{r}
decon_mtrx_ls <- "{an_breast_10x}/{robj_dir}/decon_mtrx_breast_cancer_10x_atlas_{sample_id}_{spotlight_id}_{clust_vr}.rds" %>%
  glue::glue() %>%
  here::here() %>%
  readRDS(file = .)

# decon_mtrx_ls <- readRDS(file = sprintf("%s/%s/decon_mtrx_breast_cancer_10x_atlas_%s.rds",
#                                         an_breast_10x, robj_dir, spotlight_id))
# decon_mtrx_ls <- readRDS(file = sprintf("%s/%s/decon_mtrx_breast_cancer_10x_pdac_atlas.rds", an_breast_10x, robj_dir))
# decon_mtrx_ls <- readRDS(file = sprintf("%s/%s/decon_mtrx_breast_cancer_10x_atlas_%s_reg.rds",
#                        an_breast_10x, robj_dir, spotlight_id))
# 
# decon_mtrx_ls <- readRDS(file = sprintf("%s/%s/decon_mtrx_breast_cancer_10x_atlas_%s_latent.rds",
#                        an_breast_10x, robj_dir, spotlight_id))

decon_mtrx <- decon_mtrx_ls[[2]]
decon_mtrx <- decon_mtrx[, colnames(decon_mtrx) != "res_ss"]

# Set as 0 cell types predicted to be under 3 % of the spot
decon_mtrx[decon_mtrx < 0.03] <- 0
```

Change names to original ones and reorder according to Paula
```{r}
new_colnames <- data.frame(ct_name = colnames(decon_mtrx), stringsAsFactors = FALSE) %>%
  dplyr::left_join(col_df, by = "ct_name") %>%
  dplyr::pull(plt_name)

colnames(decon_mtrx) <- new_colnames
```

Add deconvolution matrix to Seurat object metadata
```{r}
se_obj@meta.data <- cbind(se_obj@meta.data, round(decon_mtrx, 3))
```

### Visualization
Before we start we take a look at the topic profiles
```{r}
# Extract coefficient matrix
h <- NMF::coef(decon_mtrx_ls[[1]][[1]])

# Extract coefficient labels
train_labs <- data.frame(ct_name = decon_mtrx_ls[[1]][[2]]) %>%
  dplyr::left_join(col_df, by = "ct_name") %>%
  dplyr::pull(plt_name)

# rownames(h) <- paste("Topic", 1:nrow(h), sep = " ")
profile_plt <- SPOTlight::dot_plot_profiles_fun(
  h = h,
  train_cell_clust = train_labs)

plt_2 <- profile_plt[[2]] +
  ggplot2::scale_x_discrete(limits = unique(train_labs)) +
  ggplot2::scale_y_discrete(labels = glue::glue("Topic {1:nrow(h)}")) +
  ggplot2::labs(x = "", y = "") +
  ggplot2::theme(
    axis.title = ggplot2::element_text(size = 15),
    axis.text.x = ggplot2::element_text(hjust = 1)
    )
profile_plt[[1]]

"{an_breast_10x}/{plt_dir}/all_ct_profiles_{spotlight_id}_{sample_id}_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = profile_plt[[1]],
    base_height = 30,
    base_width = 30)

"{an_breast_10x}/{plt_dir}/ct_profiles_{spotlight_id}_{sample_id}_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = plt_2 +
      ggplot2::theme(axis.title = ggplot2::element_text(size = 18)),
    base_height = 15,
    base_width = 15)
```

Lastly we can take a look at which genes are the most important for each topic and get an insight into which genes are driving them.
```{r}
basis_spotlight <- data.frame(NMF::basis(decon_mtrx_ls[[1]][[1]]))

# colnames(basis_spotlight) <- unique(stringr::str_wrap(decon_mtrx_ls[[1]][[2]], width = 30))
colnames(basis_spotlight) <- glue::glue("Topic-{1:length(unique(decon_mtrx_ls[[1]][[2]]))}")

basis_spotlight %>%
  round(., 5) %>% 
  DT::datatable(., filter = "top")
```

We will start by looking at the location of each cell type
```{r}
# https://stackoverflow.com/questions/38722202/how-do-i-change-the-number-of-decimal-places-on-axis-labels-in-ggplot2
# Our transformation function to ensure 2 decimals in the legend
scaleFUN <- function(x) sprintf("%.2f", x)

ct_all <- colnames(decon_mtrx)
ct_plt_ls <- lapply(ct_all, function(ct) {
  tmp_plt <- Seurat::SpatialPlot(
    object = se_obj,
    features = ct,
    alpha = c(0, 1),
    # Remove background image
    image.alpha = 0)
  
  if (sum(se_obj@meta.data[, ct]) == 0) {
    tmp_plt <- tmp_plt + ggplot2::scale_alpha(range = c(0,0))
  } else {
    tmp_plt <- tmp_plt + ggplot2::scale_alpha(range = c(0, 1))
  }
  
  tmp_plt <- tmp_plt +
    ggplot2::theme(
      legend.title = ggplot2::element_blank()
      ) +
    ggplot2::labs(title = ct) +
    ggplot2::scale_fill_gradientn(
      colors = grDevices::heat.colors(10, rev = TRUE),
      # Same number of breaks for all plots
      breaks = seq(min(se_obj@meta.data[, ct]),
                   max(se_obj@meta.data[, ct]),
                   length.out = 4),
      # 2 decimals in the legend
      labels = scaleFUN
      # limits = c(0, 1)
      )
return(tmp_plt)
})

ct_grid <- cowplot::plot_grid(
  plotlist = ct_plt_ls,
  axis = "trbl",
  align = "hv",
  nrow = 5,
  ncol = 5)

"{an_breast_10x}/{plt_dir}/{sample_id}_10x_immune_reference_arrangement_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = ct_grid,
    base_height = 25,
    base_width = 25)
```

Next we will only plot the location of the cell types of interest for the main plot
```{r eval = FALSE}
# i <- unique(se_obj$slice)[1]
# ct <- cell_types[[1]]
ct_plt_ls <- lapply(ct_interest, function(ct) {
  print(ct)
  tmp_plt <- Seurat::SpatialPlot(
    object = se_obj,
    features = ct,
    alpha = c(0, 1),
    # Remove background image
    image.alpha = 0) +
    theme(legend.title = ggplot2::element_blank())
  
  if (sum(se_obj@meta.data[, ct]) == 0) {
    tmp_plt <- tmp_plt + ggplot2::scale_alpha(range = c(0,0))
  } else {
    tmp_plt <- tmp_plt +
    ggplot2::scale_fill_gradientn(
      colors = grDevices::heat.colors(10, rev = TRUE),
      # Same number of breaks for all plots
      breaks = seq(min(se_obj@meta.data[, ct]),
                   max(se_obj@meta.data[, ct]),
                   length.out = 4),
      # 2 decimals in the legend
      labels = scaleFUN
      # limits = c(0, 1)
      )
  }
  
  tmp_plt <- tmp_plt +
    ggplot2::labs(title = ct) +
    ggplot2::theme(
      plot.title = ggplot2::element_text(hjust = 0.5, size = 20, face = "bold"),
      legend.title = ggplot2::element_blank())

return(tmp_plt)
})

interest_grid <- cowplot::plot_grid(plotlist = ct_plt_ls,
                   axis = "trbl",
                   align = "hv",
                   nrow = 3,
                   ncol = 3)

"{an_breast_10x}/{plt_dir}/{sample_id}_10x_immune_reference_arrangement_interest_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = interest_grid,
    base_height = 18,
    base_width = 15)
```

Then we will plot the spatial-scatterpie for both slices to get a gist of what is going on
```{r eval = FALSE}
spsct_plt1 <- SPOTlight::spatial_scatterpie(
  se_obj = se_obj,
  cell_types_all = cell_types,
  img_path = here::here("data/breast_visium/section_1/spatial/tissue_lowres_image.pdf"),
  pie_scale = 0.4,
  slice = sample_id)

if (clust_vr == "lv1_annot") {
  spsct_plt1 <- spsct_plt1 +
    ggplot2::scale_fill_manual(
    values = col_df[col_df$plt_name %in% cell_types, "ct_col"],
    breaks = cell_types) +
    ggplot2::guides(fill = ggplot2::guide_legend(ncol = 1))
}

"{an_breast_10x}/{plt_dir}/{sample_id}_10x_spatial_scatterpie_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = spsct_plt1,
    base_width = 12,
    base_height = 9)
```

We will also plot the scatterpie without the slice image underneath
```{r}
sct_plt1 <- SPOTlight::scatterpie_plot(se_obj = se_obj,
                                       cell_types_all = ct_all,
                                       pie_scale = 0.7,
                                       slice = sample_id) +
  ggplot2::scale_fill_manual(
    values = col_df[col_df$plt_name %in% ct_all, "col_df"],
    breaks = colnames(decon_mtrx)) +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::guides(fill = ggplot2::guide_legend(ncol = 1))

"{an_breast_10x}/{plt_dir}/{sample_id}_10x_scatterpie_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = sct_plt1,
    base_width = 12,
    base_height = 9)
```

We can also take a loot at the spatial scatterpie by looking at cell types which are not present throughout the entire tissue.
```{r}
# Subset cell types from metadata
metadata_subset <- se_obj@meta.data[, ct_all]

# Create masks
keep_0.9 <- colSums(se_obj@meta.data[, ct_all] > 0) < 0.75 * ncol(se_obj)
keep_g0 <- colSums(se_obj@meta.data[, ct_all] > 0) > 0

# Select cell types fullfiling the conditions
ct_var <- colnames(decon_mtrx)[keep_0.9 & keep_g0]

col_df <- col_df %>% dplyr::arrange(plt_name)
sct_plt_int <- SPOTlight::scatterpie_plot(se_obj = se_obj,
                                       cell_types_all = ct_var,
                                       pie_scale = 0.4) +
  ggplot2::scale_fill_manual(
    values = col_df[col_df$plt_name %in% ct_var, "col_vec"],
    breaks = col_df[col_df$plt_name %in% ct_var, "plt_name"]) +
  ggplot2::labs(fill = "") +
  ggplot2::coord_fixed(ratio = 1) +
  ggplot2::theme(legend.position = "top") +
  ggplot2::guides(fill = ggplot2::guide_legend(nrow = 3))

"{an_breast_10x}/{plt_dir}/{sample_id}_10x_scatterpie_interest_{clust_vr}.pdf" %>%
  glue::glue() %>%
  here::here() %>%
  cowplot::save_plot(
    filename = .,
    plot = sct_plt_int,
    base_width = 12,
    base_height = 9)
```

### Gene - Cell correlation
Check assumptions
```{r eval = FALSE}
lapply(c("Regulatory T cells", "Cytotoxic CD8 T cells"), function(ct) {
  # Shapiro-Wilk normality test for mpg
  print(shapiro.test(se_obj@meta.data[, ct]))
  ggpubr::ggqqplot(se_obj@meta.data[, ct])

})

lapply(c("ENTPD1", "NT5E", "TGFB1"), function(gene) {
  # Shapiro-Wilk normality test for mpg
  print(shapiro.test(se_obj@assays$SCT@data[gene, ]))
  ggpubr::ggqqplot(se_obj@assays$SCT@data[gene, ])

})
```

Next we want to check if the fibrotic marker TGFB1 with Treg and Cytotoxic CD8 T cells"
```{r eval = FALSE}
library(ggpubr)

metadata <- dplyr::bind_cols(
  se_obj@meta.data,
  data.frame(t(se_obj@assays$SCT@data[c("ENTPD1", "NT5E", "TGFB1"), ])))

cor_plts <- lapply(c("Regulatory T cells", "Cytotoxic CD8 T cells"), function(ct) {
  tmp_ls <- lapply(c("ENTPD1", "NT5E", "TGFB1"), function(gene) {
    
    tmp_plt <- ggpubr::ggscatter(data = metadata,
                      x = gene,
                      y = ct,
                      add = "reg.line") +
       ggpubr::stat_cor(method = "pearson")
    return(tmp_plt)
    })
  cowplot::plot_grid(plotlist = tmp_ls, ncol = 1)
  })
  
tmp_arr <- cowplot::plot_grid(plotlist = cor_plts,
                              nrow = 1,
                              ncol = 2)
tmp_arr
```

## Session Info
```{r}
sessionInfo()
```
